backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.2
pydantic[email]==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.9
httpx==0.27.0
beautifulsoup4==4.12.3
cryptography==43.0.0
slowapi==0.1.9
psycopg2-binary==2.9.9
apscheduler>=3.10.0
pytest==8.3.0
pytest-asyncio==0.24.0
